+++ src/Models/Assignor.php
@@ -33,7 +33,7 @@ class Assignor extends Medools\Model
 
     const FOREIGN_KEYS = [
         'person' => [
-            'aryelgois\\Medools\\Models\\Person',
+            __NAMESPACE__ . '\\Person',
             'id'
         ],
         'address' => [