+++ src/Models/Person.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * This Software is part of aryelgois/bank-interchange and is provided "as is".
+ *
+ * @see LICENSE
+ */
+
+namespace aryelgois\BankInterchange\Models;
+
+use aryelgois\Medools;
+
+/**
+ * It is someone in the real world
+ *
+ * @author Aryel Mota Góis
+ * @license MIT
+ * @link https://www.github.com/aryelgois/bank-interchange
+ */
+class Person extends Medools\Models\Person
+{
+}